Move validation into separate module
