+++ graphql_client_codegen/src/query.rs
@@ -4,6 +4,7 @@
 mod fragments;
 mod operations;
 mod selection;
+mod validation;
 
 pub(crate) use fragments::{fragment_is_recursive, ResolvedFragment};
 pub(crate) use operations::ResolvedOperation;
@@ -36,75 +37,6 @@ pub(crate) struct ResolvedFragmentId(u32);
 #[derive(Debug, Clone, Copy)]
 pub(crate) struct VariableId(u32);
 
-// TODO: put this into a validation module.
-fn selection_set_contains_type_name(
-    parent_type_id: TypeId,
-    selection_set: &[SelectionId],
-    query: &Query,
-) -> bool {
-    for id in selection_set {
-        let selection = query.get_selection(*id);
-
-        match selection {
-            Selection::Typename => return true,
-            Selection::FragmentSpread(fragment_id) => {
-                let fragment = query.get_fragment(*fragment_id);
-                if fragment.on == parent_type_id
-                    && selection_set_contains_type_name(fragment.on, &fragment.selection_set, query)
-                {
-                    return true;
-                }
-            }
-            _ => (),
-        }
-    }
-
-    false
-}
-
-fn validate_typename_presence(query: &BoundQuery<'_>) -> anyhow::Result<()> {
-    for fragment in query.query.fragments.iter() {
-        let type_id = match fragment.on {
-            id @ TypeId::Interface(_) | id @ TypeId::Union(_) => id,
-            _ => continue,
-        };
-
-        if !selection_set_contains_type_name(fragment.on, &fragment.selection_set, query.query) {
-            anyhow::bail!(
-                "The `{}` fragment uses `{}` but does not select `__typename` on it. graphql-client cannot generate code for it. Please add `__typename` to the selection.",
-                &fragment.name,
-                type_id.name(query.schema),
-            )
-        }
-    }
-
-    let union_and_interface_field_selections =
-        query
-            .query
-            .selections()
-            .filter_map(|(selection_id, selection)| match selection {
-                Selection::Field(field) => match query.schema.get_field(field.field_id).r#type.id {
-                    id @ TypeId::Interface(_) | id @ TypeId::Union(_) => {
-                        Some((selection_id, id, &field.selection_set))
-                    }
-                    _ => None,
-                },
-                _ => None,
-            });
-
-    for selection in union_and_interface_field_selections {
-        if !selection_set_contains_type_name(selection.1, selection.2, query.query) {
-            anyhow::bail!(
-                "The query uses `{path}` at `{selected_type}` but does not select `__typename` on it. graphql-client cannot generate code for it. Please add `__typename` to the selection.",
-                path = full_path_prefix(selection.0, query),
-                selected_type = selection.1.name(query.schema)
-            );
-        }
-    }
-
-    Ok(())
-}
-
 pub(crate) fn resolve(
     schema: &Schema,
     query: &graphql_parser::query::Document,
@@ -126,7 +58,7 @@ pub(crate) fn resolve(
     }
 
     // Validation: to be expanded and factored out.
-    validate_typename_presence(&BoundQuery {
+    validation::validate_typename_presence(&BoundQuery {
         query: &resolved_query,
         schema,
     })?;
